Move Pricing components
