+++ src/components/Pricing/PlanCard/collapsible-details.js
