+++ src/components/Pricing/PricingAddons/AcademyIcon.js
