+++ src/sections/Products/index.js
@@ -12,7 +12,7 @@ import cloudicon from "./images/cloud.svg";
 import layer5 from "../../assets/images/layer5/5 icon/svg/light/5-light-no-trim.svg";
 import kanvas from "../../assets/images/kanvas/icon-only/kanvas-icon-color.svg";
 import { useStyledDarkMode } from "../../theme/app/useStyledDarkMode";
-import PlanCard from "../../components/PlanCard";
+import PlanCard from "../../components/Pricing/PlanCard";
 import { Row } from "../../reusecore/Layout";
 import Reviews from "../Pricing/review-slider";
 import options from "../Pricing/generatePlans";
